Implements CAN connection in text mode and single device discovery
To make the CAN bus work with the text-based functions for the UART connection, I
added some zero-termiation and tricked with the buffer length. I also added a guard
before the devices are scanned to prevent reading from uninitialized connections.
also fixed tiny typo in the frontend which led to errors and devices were not
switched when selected from the list
This also switches to the twai.h library since can.h is deprecated