Adds ESP-OTA updates with double bank and rollback
Added a handler to the webserver and had to change the endpoint for normal
device ota as well to simplifiy and reuse. Added a new TS node with
firmware version and same versioning scheme as in the charge controller
repo.

The small change in the partition table is due to alignement, the size of a
partition should be 0x10000 or 64kbytes aligned, so we "wasted" 4kbytes before.